DefsTypesForTests/DefsVariablesForTests moved into test-rte
