+++ monticore-grammar/src/test/java/de/monticore/javalight/cocos/JavaLightCocoTest.java
@@ -7,6 +7,7 @@ import de.monticore.javalight._ast.ASTJavaLightNode;
 import de.monticore.javalight._ast.ASTJavaMethod;
 import de.monticore.javalight._cocos.JavaLightCoCoChecker;
 import de.monticore.javalight.types3.JavaLightTypeCheck3;
+import de.monticore.runtime.types3.DefsTypesForTests;
 import de.monticore.symbols.basicsymbols.BasicSymbolsMill;
 import de.monticore.symbols.oosymbols._symboltable.MethodSymbol;
 import de.monticore.testjavalight.TestJavaLightMill;
@@ -14,7 +15,6 @@ import de.monticore.testjavalight._parser.TestJavaLightParser;
 import de.monticore.testjavalight._symboltable.ITestJavaLightArtifactScope;
 import de.monticore.testjavalight._symboltable.ITestJavaLightGlobalScope;
 import de.monticore.testjavalight._symboltable.TestJavaLightArtifactScope;
-import de.monticore.types3.util.DefsTypesForTests;
 import de.se_rwth.commons.Names;
 import de.se_rwth.commons.logging.Finding;
 import de.se_rwth.commons.logging.Log;
