+++ monticore-grammar/src/test/java/de/monticore/javalight/cocos/MethodExceptionThrowsTest.java
@@ -3,16 +3,12 @@
 package de.monticore.javalight.cocos;
 
 import de.monticore.javalight._cocos.JavaLightCoCoChecker;
-import de.monticore.testjavalight.TestJavaLightMill;
-import de.monticore.testjavalight._symboltable.ITestJavaLightScope;
-import de.monticore.types.check.SymTypeExpressionFactory;
-import de.monticore.types.check.SymTypeOfObject;
-import de.monticore.types3.util.DefsTypesForTests;
+import de.monticore.runtime.types3.DefsTypesForTests;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static de.monticore.types3.util.DefsTypesForTests.oOtype;
+import static de.monticore.runtime.types3.DefsTypesForTests.oOtype;
 import static org.junit.Assert.assertTrue;
 import de.se_rwth.commons.logging.Log;
