+++ monticore-grammar/src/test/java/de/monticore/types3/AbstractTypeVisitorTest.java
@@ -13,6 +13,8 @@ import de.monticore.ocl.oclexpressions.symboltable.OCLExpressionsSymbolTableComp
 import de.monticore.ocl.setexpressions.symboltable.SetExpressionsSymbolTableCompleter;
 import de.monticore.runtime.junit.AbstractMCTest;
 import de.monticore.runtime.junit.MCAssertions;
+import de.monticore.runtime.types3.DefsTypesForTests;
+import de.monticore.runtime.types3.DefsVariablesForTests;
 import de.monticore.symbols.basicsymbols.BasicSymbolsMill;
 import de.monticore.symbols.basicsymbols._symboltable.FunctionSymbol;
 import de.monticore.symbols.basicsymbols._symboltable.IBasicSymbolsScope;
@@ -26,8 +28,6 @@ import de.monticore.types.check.types3wrapper.TypeCheck3AsISynthesize;
 import de.monticore.types.mcbasictypes.MCBasicTypesMill;
 import de.monticore.types.mcbasictypes._ast.ASTMCType;
 import de.monticore.types3.util.CombineExpressionsWithLiteralsTypeTraverserFactory;
-import de.monticore.types3.util.DefsTypesForTests;
-import de.monticore.types3.util.DefsVariablesForTests;
 import de.monticore.types3.util.MapBasedTypeCheck3;
 import de.monticore.types3.util.TypeVisitorOperatorCalculator;
 import de.monticore.types3.util.WithinScopeBasicSymbolsResolver;
@@ -44,18 +44,18 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import static de.monticore.runtime.junit.MCAssertions.assertNoFindings;
-import static de.monticore.types3.util.DefsTypesForTests._booleanSymType;
-import static de.monticore.types3.util.DefsTypesForTests._boxedListSymType;
-import static de.monticore.types3.util.DefsTypesForTests._csStudentSymType;
-import static de.monticore.types3.util.DefsTypesForTests._intSymType;
-import static de.monticore.types3.util.DefsTypesForTests._linkedListSymType;
-import static de.monticore.types3.util.DefsTypesForTests._personSymType;
-import static de.monticore.types3.util.DefsTypesForTests._studentSymType;
-import static de.monticore.types3.util.DefsTypesForTests._voidSymType;
-import static de.monticore.types3.util.DefsTypesForTests.function;
-import static de.monticore.types3.util.DefsTypesForTests.inScope;
-import static de.monticore.types3.util.DefsTypesForTests.typeVariable;
-import static de.monticore.types3.util.DefsTypesForTests.variable;
+import static de.monticore.runtime.types3.DefsTypesForTests._booleanSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._boxedListSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._csStudentSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._intSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._linkedListSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._personSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._studentSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._voidSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests.function;
+import static de.monticore.runtime.types3.DefsTypesForTests.inScope;
+import static de.monticore.runtime.types3.DefsTypesForTests.typeVariable;
+import static de.monticore.runtime.types3.DefsTypesForTests.variable;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
