+++ monticore-grammar/src/test/java/de/monticore/types3/CallGenericFunctionsTest.java
@@ -18,27 +18,27 @@ import java.util.List;
 import java.util.stream.Stream;
 
 import static de.monticore.runtime.junit.MCAssertions.assertNoFindings;
+import static de.monticore.runtime.types3.DefsTypesForTests._IntegerSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._carSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._floatSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._intSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._linkedListSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._personSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._unboxedString;
+import static de.monticore.runtime.types3.DefsTypesForTests._voidSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests.function;
+import static de.monticore.runtime.types3.DefsTypesForTests.inScope;
+import static de.monticore.runtime.types3.DefsTypesForTests.type;
+import static de.monticore.runtime.types3.DefsTypesForTests.typeVariable;
+import static de.monticore.runtime.types3.DefsTypesForTests.variable;
+import static de.monticore.runtime.types3.DefsVariablesForTests._carVarSym;
+import static de.monticore.runtime.types3.DefsVariablesForTests._personVarSym;
 import static de.monticore.types.check.SymTypeExpressionFactory.createFunction;
 import static de.monticore.types.check.SymTypeExpressionFactory.createGenerics;
 import static de.monticore.types.check.SymTypeExpressionFactory.createTypeObject;
 import static de.monticore.types.check.SymTypeExpressionFactory.createTypeVariable;
 import static de.monticore.types.check.SymTypeExpressionFactory.createWildcard;
 import static de.monticore.types.mccollectiontypes.types3.util.MCCollectionSymTypeFactory.createList;
-import static de.monticore.types3.util.DefsTypesForTests._IntegerSymType;
-import static de.monticore.types3.util.DefsTypesForTests._carSymType;
-import static de.monticore.types3.util.DefsTypesForTests._floatSymType;
-import static de.monticore.types3.util.DefsTypesForTests._intSymType;
-import static de.monticore.types3.util.DefsTypesForTests._linkedListSymType;
-import static de.monticore.types3.util.DefsTypesForTests._personSymType;
-import static de.monticore.types3.util.DefsTypesForTests._unboxedString;
-import static de.monticore.types3.util.DefsTypesForTests._voidSymType;
-import static de.monticore.types3.util.DefsTypesForTests.function;
-import static de.monticore.types3.util.DefsTypesForTests.inScope;
-import static de.monticore.types3.util.DefsTypesForTests.type;
-import static de.monticore.types3.util.DefsTypesForTests.typeVariable;
-import static de.monticore.types3.util.DefsTypesForTests.variable;
-import static de.monticore.types3.util.DefsVariablesForTests._carVarSym;
-import static de.monticore.types3.util.DefsVariablesForTests._personVarSym;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
 public class CallGenericFunctionsTest
