+++ monticore-grammar/src/test/java/de/monticore/types3/CommonExpressionTypeVisitorTest.java
@@ -37,6 +37,7 @@ import java.util.List;
 
 import static de.monticore.runtime.junit.MCAssertions.assertHasFindingsStartingWith;
 import static de.monticore.runtime.junit.MCAssertions.assertNoFindings;
+import static de.monticore.runtime.types3.DefsTypesForTests.*;
 import static de.monticore.types.check.SymTypeExpressionFactory.createGenerics;
 import static de.monticore.types.check.SymTypeExpressionFactory.createIntersection;
 import static de.monticore.types.check.SymTypeExpressionFactory.createTuple;
@@ -44,7 +45,6 @@ import static de.monticore.types.check.SymTypeExpressionFactory.createTypeArray;
 import static de.monticore.types.check.SymTypeExpressionFactory.createTypeObject;
 import static de.monticore.types.check.SymTypeExpressionFactory.createTypeVariable;
 import static de.monticore.types.check.SymTypeExpressionFactory.createUnion;
-import static de.monticore.types3.util.DefsTypesForTests.*;
 
 public class CommonExpressionTypeVisitorTest
     extends AbstractTypeVisitorTest {
