+++ monticore-grammar/src/test/java/de/monticore/types3/CommonLiteralsTypeVisitorTest.java
@@ -4,8 +4,8 @@ package de.monticore.types3;
 import de.monticore.expressions.combineexpressionswithliterals.CombineExpressionsWithLiteralsMill;
 import de.monticore.literals.mccommonliterals.MCCommonLiteralsMill;
 import de.monticore.literals.mcliteralsbasis._ast.ASTLiteral;
+import de.monticore.runtime.types3.DefsTypesForTests;
 import de.monticore.types.check.SymTypeExpression;
-import de.monticore.types3.util.DefsTypesForTests;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
