+++ monticore-grammar/src/test/java/de/monticore/types3/ResolveWithinOOTypeTest.java
@@ -30,10 +30,10 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import static de.monticore.runtime.junit.MCAssertions.assertNoFindings;
-import static de.monticore.types3.util.DefsTypesForTests._intSymType;
-import static de.monticore.types3.util.DefsTypesForTests.inScope;
-import static de.monticore.types3.util.DefsTypesForTests.method;
-import static de.monticore.types3.util.DefsTypesForTests.oOtype;
+import static de.monticore.runtime.types3.DefsTypesForTests._intSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests.inScope;
+import static de.monticore.runtime.types3.DefsTypesForTests.method;
+import static de.monticore.runtime.types3.DefsTypesForTests.oOtype;
 
 /**
  * tests whether we can resolve correctly constructors within a type.
