+++ monticore-grammar/src/test/java/de/monticore/types3/SetExpressionsTypeVisitorTest.java
@@ -1,12 +1,12 @@
 /* (c) https://github.com/MontiCore/monticore */
 package de.monticore.types3;
 
+import de.monticore.runtime.types3.DefsVariablesForTests;
 import de.monticore.symbols.basicsymbols.BasicSymbolsMill;
 import de.monticore.symbols.basicsymbols._symboltable.IBasicSymbolsGlobalScope;
 import de.monticore.types.check.SymTypeExpressionFactory;
 import de.monticore.types.mccollectiontypes.types3.MCCollectionSymTypeRelations;
 import de.monticore.types.mccollectiontypes.types3.util.MCCollectionSymTypeFactory;
-import de.monticore.types3.util.DefsVariablesForTests;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -16,14 +16,14 @@ import org.junit.jupiter.params.provider.MethodSource;
 import java.io.IOException;
 import java.util.stream.Stream;
 
-import static de.monticore.types3.util.DefsTypesForTests._BooleanSymType;
-import static de.monticore.types3.util.DefsTypesForTests._booleanSymType;
-import static de.monticore.types3.util.DefsTypesForTests._unboxedListSymType;
-import static de.monticore.types3.util.DefsTypesForTests._unboxedSetSymType;
-import static de.monticore.types3.util.DefsTypesForTests.inScope;
-import static de.monticore.types3.util.DefsTypesForTests.variable;
-import static de.monticore.types3.util.DefsVariablesForTests._intUnboxedListVarSym;
-import static de.monticore.types3.util.DefsVariablesForTests._intUnboxedSetVarSym;
+import static de.monticore.runtime.types3.DefsTypesForTests._BooleanSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._booleanSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._unboxedListSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._unboxedSetSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests.inScope;
+import static de.monticore.runtime.types3.DefsTypesForTests.variable;
+import static de.monticore.runtime.types3.DefsVariablesForTests._intUnboxedListVarSym;
+import static de.monticore.runtime.types3.DefsVariablesForTests._intUnboxedSetVarSym;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
 public class SetExpressionsTypeVisitorTest extends AbstractTypeVisitorTest {
