+++ monticore-grammar/src/test/java/de/monticore/types3/StreamExpressionsTypeVisitorTest.java
@@ -1,8 +1,8 @@
 /* (c) https://github.com/MontiCore/monticore */
 package de.monticore.types3;
 
+import de.monticore.runtime.types3.DefsVariablesForTests;
 import de.monticore.types3.streams.StreamSymTypeRelations;
-import de.monticore.types3.util.DefsVariablesForTests;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
