+++ monticore-grammar/src/test/java/de/monticore/types3/SymTypeLeastUpperBoundTest.java
@@ -5,8 +5,8 @@ package de.monticore.types3;
 import de.monticore.expressions.combineexpressionswithliterals.CombineExpressionsWithLiteralsMill;
 import de.monticore.expressions.combineexpressionswithliterals._symboltable.ICombineExpressionsWithLiteralsScope;
 import de.monticore.runtime.junit.AbstractMCTest;
+import de.monticore.runtime.types3.DefsTypesForTests;
 import de.monticore.types.check.SymTypeExpression;
-import de.monticore.types3.util.DefsTypesForTests;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -14,16 +14,15 @@ import org.junit.jupiter.api.Test;
 import java.util.Optional;
 
 import static de.monticore.runtime.junit.MCAssertions.assertNoFindings;
+import static de.monticore.runtime.types3.DefsTypesForTests._carSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._childSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._csStudentSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._personSymType;
+import static de.monticore.runtime.types3.DefsTypesForTests._studentSymType;
 import static de.monticore.types.check.SymTypeExpressionFactory.createBottomType;
 import static de.monticore.types.check.SymTypeExpressionFactory.createIntersection;
 import static de.monticore.types.check.SymTypeExpressionFactory.createTypeArray;
 import static de.monticore.types.check.SymTypeExpressionFactory.createUnion;
-import static de.monticore.types3.util.DefsTypesForTests._carSymType;
-import static de.monticore.types3.util.DefsTypesForTests._childSymType;
-import static de.monticore.types3.util.DefsTypesForTests._csStudentSymType;
-import static de.monticore.types3.util.DefsTypesForTests._personSymType;
-import static de.monticore.types3.util.DefsTypesForTests._studentSymType;
-import static org.junit.Assert.assertEquals;
 
 public class SymTypeLeastUpperBoundTest extends AbstractMCTest {
